Mock rails more sensibly for RSpec
